Overflow error with tokio
